Fix data load & cleanup
하드코딩되어있던 부분을 조금 수정했습니다. & 불러오는 로직을 재사용성을 높이기위해 수정했습니다.
